Revert #571 and move the test to test_all instead
We don't run test_all unfiltered against TARGET=ibs-cr so this should be fine 🤞
